+++ priv/install/templates/static/villain/villain.all.js
@@ -3732,6 +3732,8 @@ if (typeof exports === 'object') {
             'mouseout .villain-block-inner': 'onMouseOut',
             'paste .villain-text-block': 'onPaste',
             'mouseup .villain-text-block': 'onMouseUp',
+            'keyup .villain-text-block': 'onKeyUp',
+    
             'click .villain-text-block': 'onClick',
             'click .villain-action-button-setup': 'onSetupClick'
         },
@@ -3783,6 +3785,17 @@ if (typeof exports === 'object') {
             } else {
                 $button.addClass('active');
                 this.showSetup();
+            }
+        },
+    
+        onKeyUp: function(e) {
+            // check if there's text selected
+            var text = this.getSelectedText();
+    
+            if (text !== '') {
+                Villain.EventBus.trigger('formatpopup:show', this);
+            } else {
+                Villain.EventBus.trigger('formatpopup:hide');
             }
         },